move base packet classes to a base file
